+++ projects/angular-cesium/src/lib/angular-cesium-widgets/services/entity-editors/polygons-editor/polygons-editor.service.ts
@@ -17,9 +17,9 @@ import { PolygonsManagerService } from './polygons-manager.service';
 import { PolygonEditorObservable } from '../../../models/polygon-editor-observable';
 import { EditablePolygon } from '../../../models/editable-polygon';
 import { PolygonEditOptions, PolygonProps } from '../../../models/polygon-edit-options';
-import { PointProps } from '../../../models/polyline-edit-options';
+import { ClampTo3DOptions, PointProps } from '../../../models/polyline-edit-options';
 import { LabelProps } from '../../../models/label-props';
-import { generateKey } from '../../utils';
+import { debounce, generateKey } from '../../utils';
 
 export const DEFAULT_POLYGON_OPTIONS: PolygonEditOptions = {
   addPointEvent: CesiumEvent.LEFT_CLICK,
@@ -54,6 +54,8 @@ export const DEFAULT_POLYGON_OPTIONS: PolygonEditOptions = {
   clampHeightTo3D: false,
   clampHeightTo3DOptions: {
     clampToTerrain: false,
+    clampMostDetailed: true,
+    clampToHeightPickWidth: 2,
   },
 };
 
@@ -101,6 +103,10 @@ export class PolygonsEditorService {
   private observablesMap = new Map<string, DisposableObservable<any>[]>();
   private cesiumScene: any;
 
+  private clampPointsDebounced = debounce((id, clampHeightTo3D: boolean, clampHeightTo3DOptions) => {
+    this.clampPoints(id, clampHeightTo3D, clampHeightTo3DOptions);
+  }, 300);
+
   init(mapEventsManager: MapEventsManagerService,
        coordinateConverter: CoordinateConverter,
        cameraService: CameraService,
@@ -120,7 +126,36 @@ export class PolygonsEditorService {
     return this.updatePublisher;
   }
 
-  screenToPosition(cartesian2, clampHeightTo3D: boolean, isTerrain?) {
+  private clampPoints(id, clampHeightTo3D: boolean, { clampToTerrain, clampMostDetailed, clampToHeightPickWidth }: ClampTo3DOptions) {
+    if (clampHeightTo3D && clampMostDetailed) {
+      const polygon = this.polygonsManager.get(id);
+      const points = polygon.getPoints();
+
+      if (!clampToTerrain) {
+        // 3dTiles
+        points.forEach(point => {
+          point.setPosition(this.cesiumScene.clampToHeight(point.getPosition(), undefined, clampToHeightPickWidth));
+        });
+        // const cartesians = points.map(point => point.getPosition());
+        // const promise = this.cesiumScene.clampToHeightMostDetailed(cartesians, undefined, clampToHeightPickWidth);
+        // promise.then((updatedCartesians) => {
+        //   points.forEach((point, index) => {
+        //     point.setPosition(updatedCartesians[index]);
+        //   });
+        // });
+      } else {
+        const cartographics = points.map(point => this.coordinateConverter.cartesian3ToCartographic(point.getPosition()));
+        const promise = Cesium.sampleTerrain(this.cesiumScene.terrainProvider, 11, cartographics);
+        Cesium.when(promise, (updatedPositions) => {
+          points.forEach((point, index) => {
+            point.setPosition(Cesium.Cartographic.toCartesian(updatedPositions[index]));
+          });
+        });
+      }
+    }
+  }
+
+  private screenToPosition(cartesian2, clampHeightTo3D: boolean, { clampToHeightPickWidth, clampToTerrain }: ClampTo3DOptions) {
     const cartesian3 = this.coordinateConverter.screenToCartesian3(cartesian2);
 
     // If cartesian3 is undefined then the point inst on the globe
@@ -130,7 +165,8 @@ export class PolygonsEditorService {
         return this.cesiumScene.globe.pick(ray, this.cesiumScene);
       };
 
-      if (isTerrain) {
+      // is terrain?
+      if (clampToTerrain) {
         return globePositionPick();
       } else {
         const cartesian3PickPosition = this.cesiumScene.pickPosition(cartesian2);
@@ -138,7 +174,7 @@ export class PolygonsEditorService {
         if (latLon.height < 0) {// means nothing picked -> Validate it
           return globePositionPick();
         }
-        return this.cesiumScene.clampToHeight(cartesian3PickPosition);
+        return this.cesiumScene.clampToHeight(cartesian3PickPosition, undefined, clampToHeightPickWidth);
       }
     }
 
@@ -168,18 +204,21 @@ export class PolygonsEditorService {
     const mouseMoveRegistration = this.mapEventsManager.register({
       event: CesiumEvent.MOUSE_MOVE,
       pick: PickOptions.NO_PICK,
+      pickConfig: options.pickConfiguration,
       priority,
     });
     const addPointRegistration = this.mapEventsManager.register({
       event: polygonOptions.addPointEvent,
       modifier: polygonOptions.addPointModifier,
       pick: PickOptions.NO_PICK,
+      pickConfig: options.pickConfiguration,
       priority,
     });
     const addLastPointRegistration = this.mapEventsManager.register({
       event: polygonOptions.addLastPointEvent,
       modifier: polygonOptions.addLastPointModifier,
       pick: PickOptions.NO_PICK,
+      pickConfig: options.pickConfiguration,
       priority,
     });
 
@@ -187,8 +226,7 @@ export class PolygonsEditorService {
     const editorObservable = this.createEditorObservable(clientEditSubject, id);
 
     mouseMoveRegistration.subscribe(({ movement: { endPosition } }) => {
-      const position = this.screenToPosition(endPosition, polygonOptions.clampHeightTo3D,
-        polygonOptions.clampHeightTo3DOptions.clampToTerrain);
+      const position = this.screenToPosition(endPosition, polygonOptions.clampHeightTo3D, polygonOptions.clampHeightTo3DOptions);
 
       if (position) {
         this.updateSubject.next({
@@ -205,8 +243,7 @@ export class PolygonsEditorService {
       if (finishedCreate) {
         return;
       }
-      const position = this.screenToPosition(endPosition, polygonOptions.clampHeightTo3D,
-          polygonOptions.clampHeightTo3DOptions.clampToTerrain);
+      const position = this.screenToPosition(endPosition, polygonOptions.clampHeightTo3D, polygonOptions.clampHeightTo3DOptions);
       if (!position) {
         return;
       }
@@ -244,8 +281,7 @@ export class PolygonsEditorService {
 
 
     addLastPointRegistration.subscribe(({ movement: { endPosition } }) => {
-      const position = this.screenToPosition(endPosition, polygonOptions.clampHeightTo3D,
-        polygonOptions.clampHeightTo3DOptions.clampToTerrain);
+      const position = this.screenToPosition(endPosition, polygonOptions.clampHeightTo3D, polygonOptions.clampHeightTo3DOptions);
       if (!position) {
         return;
       }
@@ -341,11 +377,13 @@ export class PolygonsEditorService {
                       editSubject: Subject<PolygonEditUpdate>,
                       options: PolygonEditOptions,
                       editObservable?: PolygonEditorObservable): PolygonEditorObservable {
+    this.clampPoints(id, options.clampHeightTo3D, options.clampHeightTo3DOptions);
 
     const pointDragRegistration = this.mapEventsManager.register({
       event: options.dragPointEvent,
       entityType: EditPoint,
       pick: PickOptions.PICK_FIRST,
+      pickConfig: options.pickConfiguration,
       priority,
       pickFilter: entity => id === entity.editedEntityId,
     });
@@ -356,6 +394,7 @@ export class PolygonsEditorService {
         event: options.dragShapeEvent,
         entityType: EditablePolygon,
         pick: PickOptions.PICK_FIRST,
+        pickConfig: options.pickConfiguration,
         priority,
         pickFilter: entity => id === entity.id,
       });
@@ -365,6 +404,7 @@ export class PolygonsEditorService {
       entityType: EditPoint,
       modifier: options.removePointModifier,
       pick: PickOptions.PICK_FIRST,
+      pickConfig: options.pickConfiguration,
       priority,
       pickFilter: entity => id === entity.editedEntityId,
     });
@@ -372,7 +412,7 @@ export class PolygonsEditorService {
     pointDragRegistration.pipe(
       tap(({ movement: { drop } }) => this.cameraService.enableInputs(drop)))
       .subscribe(({ movement: { endPosition, drop }, entities }) => {
-        const position = this.screenToPosition(endPosition, options.clampHeightTo3D, options.clampHeightTo3DOptions.clampToTerrain);
+        const position = this.screenToPosition(endPosition, options.clampHeightTo3D, options.clampHeightTo3DOptions);
         if (!position) {
           return;
         }
@@ -392,14 +432,16 @@ export class PolygonsEditorService {
           positions: this.getPositions(id),
           points: this.getPoints(id),
         });
+
+        this.clampPointsDebounced(id, options.clampHeightTo3D, options.clampHeightTo3DOptions);
       });
 
     if (shapeDragRegistration) {
       shapeDragRegistration
         .pipe(tap(({ movement: { drop } }) => this.cameraService.enableInputs(drop)))
         .subscribe(({ movement: { startPosition, endPosition, drop }, entities }) => {
-          const endDragPosition = this.screenToPosition(endPosition, false);
-          const startDragPosition = this.screenToPosition(startPosition, false);
+          const endDragPosition = this.screenToPosition(endPosition, false, options.clampHeightTo3DOptions);
+          const startDragPosition = this.screenToPosition(startPosition, false, options.clampHeightTo3DOptions);
           if (!endDragPosition) {
             return;
           }
@@ -445,6 +487,8 @@ export class PolygonsEditorService {
         positions: this.getPositions(id),
         points: this.getPoints(id),
       });
+
+      this.clampPoints(id, options.clampHeightTo3D, options.clampHeightTo3DOptions);
     });
 
     const observables = [pointDragRegistration, pointRemoveRegistration];
@@ -465,9 +509,10 @@ export class PolygonsEditorService {
 
     const defaultClone = JSON.parse(JSON.stringify(DEFAULT_POLYGON_OPTIONS));
     const polygonOptions: PolygonEditOptions = Object.assign(defaultClone, options);
-    polygonOptions.pointProps = Object.assign({}, DEFAULT_POLYGON_OPTIONS.pointProps, options.pointProps);
-    polygonOptions.polygonProps = Object.assign({}, DEFAULT_POLYGON_OPTIONS.polygonProps, options.polygonProps);
-    polygonOptions.polylineProps = Object.assign({}, DEFAULT_POLYGON_OPTIONS.polylineProps, options.polylineProps);
+    polygonOptions.pointProps = { ...DEFAULT_POLYGON_OPTIONS.pointProps, ...options.pointProps};
+    polygonOptions.polygonProps = {...DEFAULT_POLYGON_OPTIONS.polygonProps, ...options.polygonProps};
+    polygonOptions.polylineProps = {...DEFAULT_POLYGON_OPTIONS.polylineProps, ...options.polylineProps};
+    polygonOptions.clampHeightTo3DOptions = { ...DEFAULT_POLYGON_OPTIONS.clampHeightTo3DOptions, ...options.clampHeightTo3DOptions};
 
     if (options.clampHeightTo3D) {
       if (!this.cesiumScene.pickPositionSupported || !this.cesiumScene.clampToHeightSupported) {