+++ projects/angular-cesium/src/lib/angular-cesium-widgets/services/utils.ts
@@ -7,3 +7,12 @@ export function generateKey(length = 12): string {
 
   return id;
 }
+
+export function debounce(func, wait) {
+  let timeout;
+  return function (...args) {
+    const context = this;
+    clearTimeout(timeout);
+    timeout = setTimeout(() => func.apply(context, args), wait);
+  };
+}