+++ projects/angular-cesium/src/lib/angular-cesium/services/map-events-mananger/event-registration-input.ts
@@ -2,6 +2,12 @@ import { CesiumEvent } from './consts/cesium-event.enum';
 import { CesiumEventModifier } from './consts/cesium-event-modifier.enum';
 import { PickOptions } from './consts/pickOptions.enum';
 
+export interface PickConfiguration {
+  pickHeight?: number;
+  pickWidth?: number;
+  drillPickLimit?: number;
+}
+
 /**
  * Interface for Event Registration Input
  * __usage:__
@@ -16,4 +22,5 @@ export interface EventRegistrationInput {
   priority?: number;
   pick?: PickOptions;
   pickFilter?: (any) => boolean;
+  pickConfig?: PickConfiguration;
 }